publish_qc: refactor to move the functionality into its own module.
